variance value long to float

+++ main.cpp
@@ -167,8 +167,8 @@ int main(int argc, char *argv[]) {
     #else
         printf("End of iterations\n"); fflush(stdout);
     #endif
-        long inVariance = computeClusterVariance(data, pixelsMap, numPixels, centroids, numBands);
-        printf("(k=%d) Within-class variance: %d\n\n", numClusters, inVariance); fflush(stdout);
+        float inVariance = computeClusterVariance(data, pixelsMap, numPixels, centroids, numBands);
+        printf("(k=%d) Within-class variance: %f\n\n", numClusters, inVariance); fflush(stdout);
 
     #ifdef USE_OMP
     #pragma omp critical
